src/mcp_client/requirements.txt
fastapi==0.118.2
mcp==1.16.0
openai==2.3.0
pydantic==2.12.0
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
uvicorn==0.37.0
